requirements.txt
numpy==1.22.2
pandas==1.4.0
Pillow==9.4.0
plotly==5.6.0
streamlit==1.19.0
streamlit_option_menu==0.3.2
sklearn==0.0
bcrypt==4.0.0
requests==2.27.1